% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imc.R
\name{imc}
\alias{imc}
\title{Indice de Masse Musculaire}
\usage{
imc(masse, taille)
}
\arguments{
\item{masse}{valeur numerique en kg superieure a 0}

\item{taille}{valeur numerique en m superieure a 0}
}
\value{
la valeur de imc
}
\description{
Indice de Masse Musculaire
}
\examples{
imc(masse = 80, taille = 1.80)
}
